Move config/sds/theme to config/theme for simplicity
