+++ src/navigators/TabNavigator.tsx
@@ -4,7 +4,7 @@ import { HomeScreen } from "components/screens/HomeScreen";
 import { SettingsScreen } from "components/screens/SettingsScreen";
 import { SwapScreen } from "components/screens/SwapScreen";
 import { ROUTES, TabStackParamList } from "config/routes";
-import { THEME } from "config/sds/theme";
+import { THEME } from "config/theme";
 import { px } from "helpers/dimensions";
 import React from "react";
 import styled from "styled-components/native";
